requirements.txt
pandas==2.2.2
sqlalchemy==1.4.54
pymysql==1.1.1
mysql-connector-python==9.0.0
apache-airflow==2.9.3
